requirements.txt
langchain-community
langgraph
pydantic
langchain-core
python-dotenv
requests
langchain-nvidia-ai-endpoints
streamlit
Pillow

